Moves formats to fs_implementation and search_authors to formating
Deletes the empty services/search directory
Deletes the empty services/util directory
