+++ browse/controllers/tb_page/__init__.py
@@ -17,7 +17,7 @@ from browse.domain.identifier import Identifier, IdentifierException
 from browse.services.documents import get_doc_service
 from browse.services.documents.base_documents import AbsException, \
     AbsNotFoundException
-from browse.services.search.search_authors import queries_for_authors, \
+from browse.formating.search_authors import queries_for_authors, \
     split_long_author_list
 
 app_config = get_application_config()
