--- browse/services/search/__init__.py
@@ -1,1 +1,0 @@
-"""Author search utility."""
